https://github.com/git-commit-id/maven-git-commit-id-plugin/issues/228: move another set of classes into the core module
