+++ core/src/main/java/pl/project13/core/JGitProvider.java
@@ -15,7 +15,7 @@
  * along with git-commit-id-plugin.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package pl.project13.maven.git;
+package pl.project13.core;
 
 import com.google.common.annotations.VisibleForTesting;
 
@@ -28,8 +28,6 @@ import org.eclipse.jgit.revwalk.RevWalk;
 import org.eclipse.jgit.revwalk.RevWalkUtils;
 import org.eclipse.jgit.storage.file.FileRepositoryBuilder;
 
-import pl.project13.core.AheadBehind;
-import pl.project13.core.GitCommitIdExecutionException;
 import pl.project13.core.jgit.DescribeResult;
 import pl.project13.core.jgit.JGitCommon;
 import pl.project13.core.jgit.DescribeCommand;
