+++ core/src/main/java/pl/project13/core/cibuild/GitlabBuildServerData.java
@@ -15,50 +15,46 @@
  * along with git-commit-id-plugin.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package pl.project13.maven.git.build;
+package pl.project13.core.cibuild;
 
 import pl.project13.core.GitCommitPropertyConstant;
 import pl.project13.core.log.LoggerBridge;
 
 import javax.annotation.Nonnull;
-import java.util.*;
+import java.util.Map;
+import java.util.Properties;
 
-public class BambooBuildServerData extends BuildServerDataProvider {
+public class GitlabBuildServerData extends BuildServerDataProvider {
 
-  BambooBuildServerData(LoggerBridge log, @Nonnull Map<String, String> env) {
-    super(log, env);
+  GitlabBuildServerData(LoggerBridge log, @Nonnull Map<String, String> env) {
+    super(log,env);
   }
 
+  /**
+   * @see <a href="https://docs.gitlab.com/ce/ci/variables/#predefined-variables-environment-variables">GitlabCIVariables</a>
+   */
   public static boolean isActiveServer(Map<String, String> env) {
-    return env.containsKey("bamboo_buildKey") ||
-            env.containsKey("bamboo.buildKey") ||
-            env.containsKey("BAMBOO_BUILDKEY");
+    return env.containsKey("CI");
   }
 
   @Override
   void loadBuildNumber(@Nonnull Properties properties) {
-    String buildNumber = Optional.ofNullable(
-            env.get("bamboo.buildNumber")).orElseGet(() -> env.get("BAMBOO_BUILDNUMBER"));
+    // GITLAB CI
+    // CI_PIPELINE_ID will be present if in a Gitlab CI environment (Gitlab >8.10 & Gitlab CI >0.5)  and contains a server wide unique ID for a pipeline run
+    String uniqueBuildNumber = env.get("CI_PIPELINE_ID");
+    // CI_PIPELINE_IID will be present if in a Gitlab CI environment (Gitlab >11.0) and contains the project specific build number
+    String buildNumber = env.get("CI_PIPELINE_IID");
 
     put(properties, GitCommitPropertyConstant.BUILD_NUMBER, buildNumber == null ? "" : buildNumber);
+    put(properties,
+        GitCommitPropertyConstant.BUILD_NUMBER_UNIQUE,
+        uniqueBuildNumber == null ? "" : uniqueBuildNumber);
   }
 
   @Override
   public String getBuildBranch() {
-    String environmentBasedKey = null;
-    String environmentBasedBranch = null;
-
-    for (String envKey : Arrays.asList(
-            "bamboo.planRepository.branchName",
-            "bamboo.planRepository.<position>.branchName",
-            "BAMBOO_PLANREPOSITORY_BRANCH")) {
-      environmentBasedBranch = env.get(envKey);
-      if (environmentBasedBranch != null) {
-        environmentBasedKey = envKey;
-        break;
-      }
-    }
-    log.info("Using environment variable based branch name. {} = {}", environmentBasedKey, environmentBasedBranch);
+    String environmentBasedBranch = env.get("CI_COMMIT_REF_NAME");
+    log.info("Using environment variable based branch name. CI_COMMIT_REF_NAME = {}", environmentBasedBranch);
     return environmentBasedBranch;
   }
 }
